wx-web-backend/requirements.txt
paho-mqtt
# Only paho-mqtt is needed for mqtt_to_db.py
